@@ -1,22 +1,20 @@
-package com.acme.edu.iteration01;
+package com.db.edu.iteration1;
 
-import com.acme.edu.Logger;
-import com.acme.edu.SysoutCaptureAndAssertionAbility;
-import org.junit.After;
-import org.junit.Before;
-import org.junit.Test;
+import com.db.edu.Logger;
+import com.db.edu.SysoutCaptureAndAssertionAbility;
+import org.junit.jupiter.api.*;
 
 import java.io.*;
 
 public class LoggerTest implements SysoutCaptureAndAssertionAbility {
     //region given
-    @Before
+    @BeforeEach
     public void setUpSystemOut() throws IOException {
         resetOut();
         captureSysout();
     }
 
-    @After
+    @AfterEach
     public void tearDown() {
         resetOut();
     }
@@ -32,16 +30,16 @@ public class LoggerTest implements SysoutCaptureAndAssertionAbility {
 
         //region then
         assertSysoutContains("primitive: ");
-        assertSysoutEquals("primitive: 1\nprimitive: 0\nprimitive: -1\n");
+        assertSysoutEquals("primitive: 1\r\nprimitive: 0\r\nprimitive: -1\r\n");
         //endregion
     }
 
     @Test
     public void shouldLogByte() throws IOException {
         //region when
-        Logger.log((byte)1);
-        Logger.log((byte)0);
-        Logger.log((byte)-1);
+        Logger.log((byte) 1);
+        Logger.log((byte) 0);
+        Logger.log((byte) -1);
         //endregion
 
         //region then
@@ -52,16 +50,13 @@ public class LoggerTest implements SysoutCaptureAndAssertionAbility {
         //endregion
     }
 
-    /*
-    TODO: implement Logger solution to match specification as tests
-
+    //TODO: implement Logger solution to match specification as tests
     @Test
     public void shouldLogChar() throws IOException {
         //region when
         Logger.log('a');
         Logger.log('b');
         //endregion
-
         //region then
         assertSysoutContains("char: ");
         assertSysoutContains("a");
@@ -75,7 +70,6 @@ public class LoggerTest implements SysoutCaptureAndAssertionAbility {
         Logger.log("test string 1");
         Logger.log("other str");
         //endregion
-
         //region then
         assertSysoutContains("string: ");
         assertSysoutContains("test string 1");
@@ -89,7 +83,6 @@ public class LoggerTest implements SysoutCaptureAndAssertionAbility {
         Logger.log(true);
         Logger.log(false);
         //endregion
-
         //region then
         assertSysoutContains("primitive: ");
         assertSysoutContains("true");
@@ -102,12 +95,9 @@ public class LoggerTest implements SysoutCaptureAndAssertionAbility {
         //region when
         Logger.log(new Object());
         //endregion
-
         //region then
         assertSysoutContains("reference: ");
         assertSysoutContains("@");
         //endregion
     }
-
-    */
 }
